requirements.txt
sounddevice
numpy
scipy
openwakeword
onnxruntime
ollama
duckduckgo-search
Pillow
fastapi
python-multipart
uvicorn
jinja2
requests
